requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==1.2.0          # via flask-restful
attrs==19.1.0             # via jsonschema
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask
concurrentloghandler==0.9.1
decorator==4.0.11         # via sqlalchemy-migrate
dominate==2.3.1           # via flask-bootstrap
flask-bootstrap==3.3.7.1
flask-httpauth==3.2.2
flask-restful==0.3.5
flask-sqlalchemy==2.2
flask==0.12.3
gevent==1.2.2
greenlet==0.4.12          # via gevent
gunicorn==19.7.1
idna==2.6                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
jsl==0.2.4
jsonschema==3.0.1
markupsafe==1.0           # via jinja2
passlib==1.7.1
pbr==2.1.0                # via sqlalchemy-migrate
pymysql==0.7.11
pyrsistent==0.15.1        # via jsonschema
python-dateutil==2.6.0
pytz==2017.2              # via flask-restful
requests==2.20.1
six==1.11.0               # via flask-restful, jsonschema, pyrsistent, python-dateutil, sqlalchemy-migrate
sqlalchemy-migrate==0.11.0
sqlalchemy==1.1.9
sqlparse==0.2.3           # via sqlalchemy-migrate
tempita==0.5.2            # via sqlalchemy-migrate
urllib3==1.22             # via requests
visitor==0.1.3            # via flask-bootstrap
werkzeug==0.14.1          # via flask
